% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_intersect.R
\name{plot_intersect}
\alias{plot_intersect}
\title{plot_intersect}
\usage{
plot_intersect(hmsis, hmpump)
}
\arguments{
\item{hmsis}{objeto criado pela funcao hmsis}

\item{hmpump}{objeto criado pela funcao hmpump}
}
\description{
plot_intersect
}
\examples{
Q_BC21_15 <- c(19.2, 18.2, 17.2, 16.0, 13.3, 9.9) / 3600
Hm_BC21_15 <- c(9, 10, 11, 12, 14, 16)
curva_bomba <- hmpump(Q_BC21_15, Hm_BC21_15)
curva_sis <- hmsis(
  hr = 7, hs = 1, dr_com = 60 / 1000, q = 15 / 3600, lr = 50, per = 148,
  ds_com = 75 / 1000, ls = 2, pes = 305
)
plot_intersect(curva_sis, curva_bomba)
}
